day4/pdcp_project3/requirements.txt
Flask==2.2.5 
pycryptodome==3.20.0 
matplotlib